4718400: Many quantities are held as signed that should be unsigned

+++ src/hotspot/share/oops/klass.hpp
@@ -549,8 +549,8 @@ protected:
   // ALL FUNCTIONS BELOW THIS POINT ARE DISPATCHED FROM AN OOP
   // These functions describe behavior for the oop not the KLASS.
 
-  // actual oop size of obj in memory
-  virtual int oop_size(oop obj) const = 0;
+  // actual oop size of obj in memory in word size.
+  virtual size_t oop_size(oop obj) const = 0;
 
   // Size of klass in word size.
   virtual int size() const = 0;
